refactor: separate overlay code
