+++ src/features/client/components/client-delete-form.tsx
@@ -1,51 +1,16 @@
 import { useTranslation } from 'react-i18next';
 
 import TrashBinIcon from '@/assets/icons/solid/trash-bin.svg?react';
-import { Button, Dialog } from '@/features/core';
+import { Button } from '@/features/core';
 import { useNavigate } from '@tanstack/react-router';
-import { overlay } from 'overlay-kit';
 
 import { Client } from '../hooks/use-client';
 import { useClientDeleteForm } from '../hooks/use-client-delete-form';
-
-export function ClientDeleteOverlay({
-  isOpen,
-  close,
-}: {
-  isOpen: boolean;
-  close: (_: boolean) => void;
-}) {
-  const { t } = useTranslation();
-
-  return (
-    <Dialog
-      isOpen={isOpen}
-      close={() => close(false)}
-      className="mx-10 min-w-75"
-    >
-      <Dialog.Header>{t('services.detail.delete.dialog.header')}</Dialog.Header>
-      <Dialog.Body>{t('services.detail.delete.dialog.body')}</Dialog.Body>
-      <Dialog.Footer>
-        <Dialog.Close className="w-full">
-          <Button variant="secondary" className="w-full">
-            {t('services.detail.delete.dialog.cancel')}
-          </Button>
-        </Dialog.Close>
-        <Button
-          variant="primary"
-          className="w-full"
-          onClick={() => close(true)}
-        >
-          {t('services.detail.delete.dialog.confirm')}
-        </Button>
-      </Dialog.Footer>
-    </Dialog>
-  );
-}
+import { onTrigger } from './client-delete-overlay';
 
 export function ClientDeleteForm({ client }: { client: Client }) {
   const { t } = useTranslation();
-  const { onSubmit } = useClientDeleteForm(client);
+  const { onSubmit } = useClientDeleteForm();
   const navigate = useNavigate({ from: '/clients/$id' });
 
   return (
@@ -59,18 +24,10 @@ export function ClientDeleteForm({ client }: { client: Client }) {
             variant="warning"
             disabled={client.deleteRequestedAt != null}
             prefixIcon={<TrashBinIcon />}
-            onClick={async () => {
-              const result = await overlay.openAsync<boolean>(
-                ({ isOpen, close }) => (
-                  <ClientDeleteOverlay isOpen={isOpen} close={close} />
-                ),
-              );
-
-              if (result) {
-                await onSubmit();
-                await navigate({ to: '/clients' });
-              }
-            }}
+            onClick={onTrigger(async () => {
+              await onSubmit(client);
+              await navigate({ to: '/clients' });
+            })}
           >
             {t('services.detail.delete.button')}
           </Button>
